# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the Exakta Laboratories turn taking application",
        "version": "1.0.0",
        "title": "API Turnero",
        "contact": {
            "email": "exakta@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": ["http"],
    "host": "localhost:4000",
    "basePath": "/api",
...
    "paths" : {
...
        "/turnstate": {
            "get": {
                "tags": ["Turn-States"],
                "summary" : "Get states of a turn",
                "description": "Get states of a turn",
                "produces": ["application/json"],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Turn-StatesRes"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 200,
                                "body": [
                                    {
                                        "_id": "618ed911116f8cee92f8d01a",
                                        "name": "espera toma",
                                        "description": "Estado que señala cuando un cliente ya fue atendido por un compañero recepsionista y esta esperando pasar a su toma.",
                                        "createdAt": "2021-11-12T21:13:53.352Z",
                                        "updatedAt": "2021-11-12T21:13:53.352Z",
                                        "__v": 0
                                    }
                                ],
                                "message": "Successful response"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 401,
                                "body": {
                                    "message": "unauthorized|TokenExpired"
                                },
                                "message": "Client error"
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 500,
                                "body": {
                                    "message": "Message error."
                                },
                                "message": "Server error"
                            }
                        }
                    }
                },
                "security":[ { "api_key":[] } ]
            },
            "post": {
                "tags": ["Turn-States"],
                "summary" : "Create a turn state",
                "description": "Create a turn state",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "role",
                        "description": "Turn state object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Turn-StatesReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/Turn-StatesRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 201,
                                "body": {
                                    "_id": "618ed911116f8cee92f8d01a",
                                    "name": "espera toma",
                                    "description": "Estado que señala cuando un cliente ya fue atendido por un compañero recepsionista y esta esperando pasar a su toma.",
                                    "createdAt": "2021-11-12T21:13:53.352Z",
                                    "updatedAt": "2021-11-12T21:13:53.352Z",
                                    "__v": 0
                                },
                                "message": "Successful response"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 401,
                                "body": {
                                    "message": "unauthorized|TokenExpired"
                                },
                                "message": "Client error"
                            }
                        }
                    },
                    "5XX": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "statusCode" : {
                                    "type": "number"
                                },
                                "body" : {
                                    "$ref": "#/definitions/InvalidRes"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "example": {
                                "statusCode": 500,
                                "body": {
                                    "message": "Message error."
                                },
                                "message": "Server error"
                            }
                        }
                    }
                },
                "security":[ { "api_key":[] } ]
            }
        },
...
    }, 
    "definitions": {
...
        "InvalidRes": {
            "type": "object",
            "properties": {
                "message":  {
                    "type": "string"
                }
            }
        },
...
        "Turn-StatesRes": {
            "type": "object",
            "properties": {
                "_id":  {
                    "type": "string"
                },
                "name":  {
                    "type": "string"
                },
                "description":  {
                    "type": "string"
                },
                "createdAt":  {
                    "type": "string"
                },
                "updatedAt":  {
                    "type": "string"
                },
                "__v":  {
                    "type": "number"
                }
            }
        },
        "Turn-StatesReq": {
            "type": "object",
            "properties": {
                "name":  {
                    "type": "string"
                },
                "description":  {
                    "type": "string"
                }
            },
            "required": [ 
                "name",
                "description"
            ],
            "example": {
                "name": "re-call",
                "description": "Estado que señala que se volvio a solicitar la presencia del cliente."
            }
        },
...
    },
    "securityDefinitions": {
        "api_key": {
            "type": "apiKey",
            "name": "auth",
            "in": "header"
        }
    }
}